Detalles varios I
Se agrega
- Atributo para el almacenamiento de la escala de preguntas ranking
- Cambio en el dise;o del detalle de pregunta, se agrega la valoracion
de preguntas simples de encuestas

Pendiente
- Visualizar en el detalle la escala seleccionada para las preguntas
ranking

+++ ProyectoSatisfaccion/src/main/java/com/satisfaccion/jpa/data/RespuestaEntity.java
@@ -24,7 +24,7 @@ public class RespuestaEntity {
 	@Column(name = "OBSERVACION")
 	private String observacion;
 	@Column(name = "VALORACION")
-	private float valoracion;
+	private int valoracion;
 	@Column(name = "USUARIOEVALUADO")
 	private String usuarioEvaluado;
 
@@ -63,7 +63,7 @@ public class RespuestaEntity {
 		return valoracion;
 	}
 
-	public void setValoracion(float valoracion) {
+	public void setValoracion(int valoracion) {
 		this.valoracion = valoracion;
 	}
